Added a lot of new features such as new Maneuvers, WIP chat server, fixed lag, fixed async (I recommend you still using non async), ftl jumps (unfinished), parties (unfinished) and a lot more! Before using this update, be sure that if you had a previous account on previous versions DETELE IT.

+++ BSGO Server/BSGO Server/Server/Catalogue/Others/Price.cs
@@ -20,7 +20,7 @@ namespace BSGO_Server
 
         public void Write(BgoProtocolWriter w)
         {
-            w.Write(items.Count);
+            w.Write((ushort)items.Count);
 
             foreach(KeyValuePair<ShipConsumableCard, float> item in items)
             {
